pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.4.2",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "pyinstaller>=6.12.0",
    "pyqt5>=5.15.11",
    "pysocks>=1.7.1",
    "python-magic>=0.4.27",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "stem>=1.8.2",
    "trafilatura>=2.0.0",
]
